Final version ... for now
It gets all the data from Stray wallet ... I filtered out the blocks before super staking. There are still a few blocks that i need to filter out , the ones where there's a transaction on the wallet in/out. The solution would be to check [0] in output if its not 0 that means that's a "bad" block. 

For now i filter those "bad" blocks in google sheets